Create README.md
This is the backend of an Event setup and organizing platform for DIGS. User creates a secure and authenticable account to use. Private data is securely stored in a remote database server. User decides level of privacy for events and events are user-generated. User event details such as event flyers are also securely saved for a limited amount of time and are self-destructive after the event's longevity expiration date which is set to 60 days after an event.